+++ src/sampletones/exporters/noise.py
@@ -47,7 +47,7 @@ class NoiseExporter(Exporter[NoiseInstruction]):
 
         return {
             FeatureKey.INITIAL_PITCH: initial_period,
-            FeatureKey.VOLUME: np.array(volumes),
-            FeatureKey.ARPEGGIO: np.array(arpeggio),
-            FeatureKey.DUTY_CYCLE: np.array(duty_cycles),
+            FeatureKey.VOLUME: np.array(volumes).astype(np.int8),
+            FeatureKey.ARPEGGIO: np.array(arpeggio).astype(np.int8),
+            FeatureKey.DUTY_CYCLE: np.array(duty_cycles).astype(np.int8),
         }
